new: expand `ActionSelector.ArgValue` to hold `uint64`
It will be used in subsequent commits to support setting
uint64 arguments for uprobes.
Note that `usdt` will still only support `uint32`.

+++ pkg/k8s/apis/cilium.io/v1alpha1/types.go
@@ -371,7 +371,7 @@ type ActionSelector struct {
 	ArgIndex uint32 `json:"argIndex"`
 	// +kubebuilder:validation:Optional
 	// An arg value for the set action
-	ArgValue uint32 `json:"argValue"`
+	ArgValue uint64 `json:"argValue"`
 	// +kubebuilder:validation:Optional
 	// An arg value for the override action, uprobe only.
 	ArgRegs []string `json:"argRegs,omitempty"`
